requirements.txt
altgraph==0.17.4
auto-py-to-exe==2.42.0 
bottle==0.12.25        
bottle-websocket==0.2.9
cffi==1.16.0           
Eel==0.16.0
future==1.0.0
gevent==24.2.1
gevent-websocket==0.10.1
glcontext==2.5.0
greenlet==3.0.3
llvmlite==0.42.0
moderngl==5.10.0
numba==0.59.0
numpy==1.26.4
packaging==23.2
pefile==2023.2.7
pycparser==2.21
pygame==2.5.2
PyGLM==2.7.1
pyinstaller==6.4.0
pyinstaller-hooks-contrib==2024.2
pyparsing==3.1.1
pywin32-ctypes==0.2.2
setuptools==69.1.1
whichcraft==0.6.1
zope.event==5.0
zope.interface==6.2
